osmApi: refactor – move to `services/overpass` folder
