+++ src/components/SearchBox/onSelectedFactory.ts
@@ -2,7 +2,7 @@ import Router, { NextRouter } from 'next/router';
 import { getApiId, getShortId, getUrlOsmId } from '../../services/helpers';
 import { addFeatureCenterToCache } from '../../services/osm/osmApi';
 import { getOverpassSource } from '../../services/mapStorage';
-import { performOverpassSearch } from '../../services/overpassSearch';
+import { performOverpassSearch } from '../../services/overpass/overpassSearch';
 import { t } from '../../services/intl';
 import { fitBounds } from './utils';
 import { getSkeleton } from './onHighlightFactory';
