+++ src/services/overpass/overpassSearch.ts
@@ -1,12 +1,18 @@
-import { LineString, OsmId, Point, GeometryCollection, Feature } from './types';
-import { getPoiClass } from './getPoiClass';
-import { getCenter } from './getCenter';
-import { fetchJson } from './fetch';
+import {
+  LineString,
+  OsmId,
+  Point,
+  GeometryCollection,
+  Feature,
+} from '../types';
+import { getPoiClass } from '../getPoiClass';
+import { getCenter } from '../getCenter';
+import { fetchJson } from '../fetch';
 import { Feature as FeatureGeojson, FeatureCollection, Polygon } from 'geojson';
-import { ASTNode } from '../components/SearchBox/queryWizard/ast';
-import { Bbox } from '../components/utils/MapStateContext';
-import { generateQuery } from '../components/SearchBox/queryWizard/generateQuery';
-import { isAstNode } from '../components/SearchBox/queryWizard/isAst';
+import { ASTNode } from '../../components/SearchBox/queryWizard/ast';
+import { Bbox } from '../../components/utils/MapStateContext';
+import { generateQuery } from '../../components/SearchBox/queryWizard/generateQuery';
+import { isAstNode } from '../../components/SearchBox/queryWizard/isAst';
 
 const getOverpassQuery = ([a, b, c, d], query: string) =>
   `[out:json][timeout:25][bbox:${[d, a, b, c]}];(${query};);out geom qt;`;
